Move askJetpack into ai asisstant block hook file. Keep it under window.askJetpack too for debugging
