+++ projects/plugins/jetpack/class.jetpack.php
@@ -663,73 +663,6 @@ class Jetpack {
 		 */
 		add_action( 'init', array( $this, 'deprecated_hooks' ) );
 
-		/**
-		 * Move me somewhere that isn't a total hack :)
-		 */
-		add_action(
-			'admin_head',
-			function () {
-				?>
-				<script>
-					window.askJetpack = ( question ) => {
-						const apiNonce = JP_CONNECTION_INITIAL_STATE.apiNonce;
-
-						async function requestToken() {
-							const request = await fetch( '/wp-json/jetpack/hack/get-openai-jwt?_cacheBuster=' + Date.now(), {
-								credentials: 'same-origin',
-								headers: {
-									'X-WP-Nonce': apiNonce,
-								}
-							} );
-
-							if ( ! request.ok ) {
-								throw new Error( "JWT request failed" );
-							}
-
-							const data = await request.json();
-							return {
-								token: data.token,
-								blogId: data.blog_id,
-							};
-						}
-
-						async function askQuestion() {
-							const { blogId, token } = await requestToken();
-
-							const url = new URL( 'https://public-api.wordpress.com/wpcom/v2/sites/' + blogId + '/jetpack-openai-query' );
-							url.searchParams.append( 'question', question );
-							url.searchParams.append( 'token', token );
-
-							const source = new EventSource( url.toString() );
-							let fullMessage = '';
-
-							source.addEventListener( 'error', ( err ) => {
-								console.error( err );
-							} );
-
-							source.addEventListener( 'message', e => {
-								if ( e.data === '[DONE]' ) {
-									source.close();
-									console.log( 'Done. Full message: ' + fullMessage );
-									return;
-								}
-
-								const data = JSON.parse( e.data );
-								const chunk = data.choices[0].delta.content;
-								if ( chunk ) {
-									fullMessage += chunk;
-									console.log( chunk );
-								}
-							} );
-						}
-
-						askQuestion().catch( err => console.error( err ) );
-					};
-				</script>
-				<?php
-			}
-		);
-
 		// Note how this runs at an earlier plugin_loaded hook intentionally to accomodate for other plugins.
 		add_action( 'plugin_loaded', array( $this, 'add_configure_hook' ), 90 );
 		add_action( 'network_plugin_loaded', array( $this, 'add_configure_hook' ), 90 );
